backend/requirements.txt
certifi==2023.11.17
chardet==3.0.4
click==8.1.7
Flask==2.2.3
Flask-Cors==3.0.10
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
python-dotenv==1.0.0
pytube==15.0.0
regex==2023.12.25
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
Werkzeug==2.2.2
zipp==3.17.0
